+++ src/main/java/com/springboot/monew/comment/controller/CommentApiDocs.java
@@ -3,4 +3,6 @@ package com.springboot.monew.comment.controller;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
 @Tag(name = "댓글", description = "댓글 관련 API")
-public interface CommentApiDocs {}
+public interface CommentApiDocs {
+  // Todo : 문서화 시 작업 예정
+}
