+++ src/main/java/com/springboot/monew/comment/controller/CommentLikeApiDocs.java
@@ -1,0 +1,5 @@
+package com.springboot.monew.comment.controller;
+
+public interface CommentLikeApiDocs {
+// Todo : 문서화 시 작업 예정
+}
